removed DeduplicationId class and its usages && moved TbMsgDeduplicationNodeTest to transform package
